requirements.txt
fastapi
uvicorn
Pillow
numpy
timm
einops
gunicorn